Implemented SVGProcessor.
How does it work?

It uses the well known equation for computing a function's fourier series.

The function is the loaded Paths specific points. It uses Batiks getPointAtLength function, which makes this project a whole lot easier to implement.
The integration is a approximation: it just adds up the itegral part. The Complex class gives functionality, to calculate the specific values.

I use arrays as temporary objects, so performance is quite good.